feature: 业务集形态改造（平台内置转对接CMDB） #622
将执行相关鉴权接口汇总到一个类中
