+++ src/backend/job-execute/service-job-execute/src/main/java/com/tencent/bk/job/execute/auth/TaskExecuteAuthService.java
@@ -29,9 +29,9 @@ import com.tencent.bk.job.common.iam.model.AuthResult;
 import com.tencent.bk.job.execute.model.ServersDTO;
 
 /**
- * 任务相关操作鉴权接口
+ * 任务执行相关操作鉴权接口
  */
-public interface TaskAuthService {
+public interface TaskExecuteAuthService {
 
     /**
      * 资源范围下快速分发文件鉴权
@@ -66,4 +66,56 @@ public interface TaskAuthService {
      */
     AuthResult authViewHistory(String username,
                                ResourceScope resourceScope);
+
+    /**
+     * 资源范围下执行脚本鉴权
+     *
+     * @param username      用户名
+     * @param resourceScope 资源范围
+     * @param scriptId      脚本ID
+     * @param scriptName    脚本名称，如果传入为空，则会调用ResourceNameQueryService查询
+     * @param servers       目标服务器
+     * @return 鉴权结果
+     */
+    AuthResult authExecuteScript(String username,
+                                 ResourceScope resourceScope,
+                                 String scriptId,
+                                 String scriptName,
+                                 ServersDTO servers);
+
+    /**
+     * 资源范围下运行执行方案鉴权
+     *
+     * @param username      用户名
+     * @param resourceScope 资源范围
+     * @param jobTemplateId 作业模板ID
+     * @param jobPlanId     执行方案ID
+     * @param jobPlanName   执行方案名称，如果传入为空，则会调用ResourceNameQueryService查询
+     * @return 鉴权结果
+     */
+    AuthResult authLaunchJobPlan(String username,
+                                 ResourceScope resourceScope,
+                                 Long jobTemplateId,
+                                 Long jobPlanId,
+                                 String jobPlanName);
+
+    /**
+     * 执行公共脚本鉴权
+     *
+     * @param username   用户名
+     * @param scriptId   公共脚本ID
+     * @param serversDTO 执行目标
+     * @return 鉴权结果
+     */
+    AuthResult authExecutePublicScript(String username,
+                                       String scriptId,
+                                       ServersDTO serversDTO);
+
+    /**
+     * 高危语句拦截记录查看鉴权
+     *
+     * @param username 用户名
+     * @return 鉴权结果
+     */
+    AuthResult authHighRiskDetectRecord(String username);
 }
